Add step to normalise image to full data range
This requires access to the entire dataset, which for a full dtaaset
will be notably slow.

For previewing (which is the purpose of this saver) this should be OK
but replaced with a more generic, read time max/min pixel calculation.
This is something the can be done by the framework and requested by a
plugin.