Fixes test file ammendment not matching checkpoint source
`_create_dataset` in `base_checkpoint_test.py` would add the value as a `np.array` with dtype of `np.int8`.
The `_create_dataset` in `checkpointing.py` sets it as a scalar `np.int16`.
After the test ammended the file the plugin_runner was no longer able to read the value