Attempt to make MedianFilter run iteratively
Define a new class CpuIterativePlugin, which potentially could be used in the
future as a class that any CPU plugin could inherit from to gain the ability to
be iterative.

Note the hardcoded value for CpuIterativePlugin._ip_fixed_iterations. In the
future there will be a way to set this via savu_config, but for now it's not
clear what the best way to do that is since it's still early days for iterative
plugins, so can leave this hardcoded for the moment.