Set in/out datasets in IterativePlugin such that process_frames() can refer to just data[0]
Due to the way that the swapping of the input and output datasets in the
IterativePlugin class was previously being done, it was causing
IterativeCcpiDenoising.process_frames() to need to refer to data[1] as well as
data[0].

This is because, on every iteration, self.parameters['in_datasets'] was a list
with two elements:
- the input Data object for the very first iteration
- the "original" or "clone" Data object (depending on the iteration number)

and so sometimes data[0] or data[1] is needed, depending on the iteration
number.

However, generalising this to make every plugin possibly iterative (in theory)
would require modifying the process_frames() method for every plugin class,
which isn't ideal.

The changes in this commit involves setting the input and output datasets for
every iteration such that self.parameters['in_datasets'] will only ever have
one Data object in it. This in turn means that the variable data[0] can still
be used in process_frames(), but now data[1] is no longer needing to be used.