Add IteratePluginGroup class to iterate over a group of plugins
First attempt at trying to make a group of plugins in a process list be
iterated over. Note that there's hardcoded values to indicate the indices of
the plugins that start and end the group of plugins to iterate over, as well as
an assumed process list of
- 1. TomoPhantomLoader
- 2. TomoPhantomArtifacts
- 3. AstraReconCpu
- 4. MedianFilter
- 5. MedianFilter
where the last two MedianFilters are the plugins attempting to be iterated over.