Enable iterating a single plugin using the generalised approach in baa8f0fe
Iterating a group of two MedianFilter plugins worked using the approach in
baa8f0fe, but it was failing to iterate just a single MedianFilter plugin.

In order for the approach to be general enough for the desired use-cases, it
needs to be able to iterate both a single plugin, and a group of consecutive
plugins.

This change enables iterating of a single MedianFilter plugin, as well as
iterating a group of two consecutive MedianFilter plugins.