Change how input datasets are set in IteratePluginGroup
This is a similar change to IteratePluginGroup as 94686a63 was for
IterativePlugin. It's been done so that the in_datasets parameter of the start
plugin in a group of plugins to iterate over is set to be a list with a single
element in it rather than a list with two elements in it (this is relevant for
when _ip_iteration >= 1, but for _ip_iteration = 0 this wasn't an issue, due to
how _ip_data_dict is set for iteration 0).