Add decorator that modifies a plugin's setup() method
In order for a plugin to function correctly at the end of an iterative loop, it
needs an extra output dataset (the cloned dataset).

The added decorator can be applied to a plugin's setup() method so that it is
able to setup the extra output dataset correctly, in the case of the plugin
being set in the process list to be at the end of an iterative loop.