Replace is_in_iterative_loop and iteration_number experiment metadata
The information regarding
- if processing is currently within an iterative loop
- if so, what iteration number is the processing currently at
is already contained in instances of IteratePluginGroup, so setting metadata in
the instance of Experiment is unnecessary duplication of information.

Instead, a list of all instances of IteratePluginGroup has been added to the
metadata in the instance of Experiment, and can be inspected to find this
information without the need for extra metadata entries.

Note that the is_end_plugin_in_iterate_group metadata entry still exists
though. Perhaps this can also be replaced/removed in the future to further
de-clutter the experiment's metadata.