Fix failing tests in checkpoint_test.py and template_loader_test.py
The current implementation of iterative plugins is hardcoded to assume a given
process list and to iterate particular plugin indices in that assumed process
list.

This in turn breaks many tests in savu_full_tests, as it is effectively
assuming that every process list that is being run in the tests is
supposed to iterate certain plugins, but not all plugins have been given
the ability to be iterative yet.

However, in the near future, there will be a means of telling Savu which
plugins in the given process list should be iterated, if any.

Therefore, temporarily commenting out the code that is forcing Savu to
assume that plugins in the process list should run iteratively, and then
running savu_full_tests once more, will be a more reliable indicator of
what breaking changes have been made in Savu for this iterative feature
that will actually be in place when running plugins non-iteratively.

Previously, tests in checkpoint_test.py and template_loader_test.py were
failing, and this was due to incorrectly omitting the linking of output hdf5
files when any non-iterative plugin had finished its processing. These
tests are now passing again with this change.