Add padding info to additional PluginData objects
Commit 27c132c7 introduced the creation of a few new PluginData
instances, which enabled the dataset-swapping that occurs for iterative
loops in the case when the pattern is changing between plugins in the
iterative loop.

The broken behaviour of iterating two MedianFilters that was mentioned
in that commit was due to the missing padding information that needs to
be set on PluginData objects.

This commit introduces a way to set this previously missing padding
info, and so iterating two MedianFilters is once again working.

However, note that iterating a single MedianFilter is still broken from
the changes in 27c132c7.