Get pattern from PluginData rather than plugin parameter
Only some plugins have a parameter for the pattern to be used for
processing, so plugin parameters are an unreliable place to depend on
for getting pattern information.

Instead, get the pattern from PluginData objects associated to a plugin
object, since PluginData objects are guaranteed to have pattern
information in them.