Move application of cloned data padding out of BaseMedianFilter
Instead of setting the padding for the the cloned dataset inside
BaseMedianFilter, it can be moved to a more general area: the
setup_extra_plugin_data_padding() function.

That way, any other filter plugin (not just plugins that inherit from
BaseMedianFilter) that is run in an iterative loop will:
- take all the padding that has been applied to the original output
  dataset
- apply it to the cloned dataset too
by using the setup_extra_plugin_data_padding() decorator.