Fix incorrect padding of start plugin's input data
In commit ca55f65b, the padding application in
setup_extra_plugin_data_padding() works for a special case (when the
input dataset of the start plugin of an iterative loop has the same
padding as the output dataset of the end plugin), but it is incorrect in
more general cases.

This change replaces that padding application with a more general
approach, and takes into account the possibility of the input dataset of
the start plugin and the output dataset of the end plugin having
different padding applied to them.