Remove dataset finalisation in BaseTomoPhantomLoader
The execution of Experiment._finalise_setup() in
BaseTomoPhantomLoader.post_process() was causing the extra output
dataset for end-plugins in iterative loops (the "clone" dataset) to not
be added in the plugin setup, and thus causing errors in process lists
that had both iterative loops and TomoPhantomLoader in them.

It appears to be unnecessary to "finalise" the setup of the Experiment
object in BaseTomoPhantomLoader, so that line has been removed. However,
it may be the case that in the future this setup-finalisation needs to
be added back in, in which case, an alternative fix would need to be
found.