Reset PluginData for input datasets of plugins in iterative loops
The processing of input datasets performed by plugins can change the
slicing/pattern of the input data, and thus can interfere with how input
data is re-used on subsequent iterations (ie, potentially causing
incorrect slicing being applied to input data).

By taking the slicing/pattern of the input data from what it was on the
previous iteration, the PluginData objects associated to the input
datasets of plugins in iterative loops can be reset to their previous
state, and this issue can be resolved.