Don't mark projection-shifts dataset as "to remove"
When using Projection2dAlignment in an iterative loop, marking the
dataset containing the shifts as "to remove" causes the associated hdf5
file to be closed, and thus an error occurs on the subsequent iteration
where the shifts hdf5 file is attempted to be accessed.

This is not an issue when "manually iterating plugins" (ie, duplicating
plugins in the process list to perform an iterative loop), and only
occurs when doing "non-manual iterating of plugins" (ie, using the
iterate command to specify an iterative loop in the configurator).

Keeping the shifts dataset (rather than marking it to be removed) fixes
this issue when doing iterative loops with the iterate command.

Note however that this causes the shifts dataset to be linked in the
final output NeXuS file. The shifts dataset is primarily for internal
Savu use and is not very useful when included in the output NeXuS file.
Therefore, in the future this change may be amended such that:
- when not in an iterative loop, the shifts dataset is removed
- when in an iterative loop, the shifts dataset is removed only at the
  end of all iterations