Shift relevant iterative loops down when a plugin is removed
Note that a separate method has been added in the `Content` class, which
now contains the logic to check iterative loops when a plugin has been
removed, where previously this logic was in `Content.remove()`.

This was done because, when plugins are moved, this also causes
`Content.remove()` to be called. Thus, this meant that moving plugins
could also trigger the removal of iterative loops, which is an undesired
effect.