Link either original or clone output dataset of iterative loop in NeXuS file
Previously when running iterative loops, both the original and clone output
dataset for the end plugin of the loop were linked in the final NeXuS file.

This is confusing for two reasons:
- only one or the other has the output of the very last iteration, and
  this depends on if the number of iterations was even or odd
- the name of the clone output dataset appears in the name of the link
  (ie, `/entry/final_result_itr_clone0/data`) in the NeXuS file, but the
  user didnt't set the name of this clone dataset anywhere (it has been
  done "under-the-hood" in the iterative code), so it appears as being a
  random link in the final NeXuS file to users

These changes link only the original or clone output dataset, but the
linkname in the NeXuS file has the name of the original in both cases,
regardless of whether the original or the clone has the output of the
very last iteration.

Ie, if the clone dataset has the output of the very last iteration, the
link will point to the clone dataset, but will have the name
`/entry/final_result_DATASET_NAME/data` rather than
`/entry/final_result_itr_clone0/data` (where `DATASET_NAME` is the name
of the output dataset that has been defined by the user in the
configurator when creating the process list).