Tweak DisplayFormatter logic that checks if a plugin is in a loop
The 'iterative' key was added to plugin dicts in 273e4725 so then
`DisplayFormatter` would know which plugins are in iterative loops, to
then display them differently in the configurator.

In order to have the 'iterative' key in the plugin dicts persist across
opening/closing process lists, it was necessary to save it in the
process list NeXuS file.

As an alternative, it turned out that it was possible to check if a
plugin was in an iterative loop from inside a `DisplayFormatter` object
by using the instance of `PluginList` in the
`DisplayFormatter.plugin_list_inst` attribute.

This approach doesn't require any extra information to be saved in the
process list NeXuS file in order for `DisplayFormatter` to know when a
plugin is in an iterative loop. Thus, it allows the display changes for
iterative loops in the configurator to persist across opening/closing
process list files without needing to save any extra information in the
process list NeXuS file.