Reset pattern of all plugin input datasets in an iterative loop
Commit 36f953e4 was supposed to fix the issue of the pattern of input
datasets potentially being changed by some plugins (such as
TomobarRecon3d).

However, that change only fixed the issue for the first input dataset of
plugins, and it didn't take into account situations that can occur with
plugins that have multiple input datasets. Namely, when a plugin with
multiple input datasets has the pattern change for any of the input
datasets that isn't the first one.

This change takes care of that more general case, by resetting the
pattern for all input datasets of a given plugin.