Correctly set `max_frames` of extra `PluginData` objects created for iterative loops
Before this change, the hardcoded setting of `max_frames` of these extra
`PluginData` objects to `'single'` was potentially causing the input
data being given to plugins in iterations > 0 to be different than
desired.

In the context of iterative alignment, this bug was causing the
`ForwardProjectorGpu` plugin to be given 3D data (as expected) on
iteration 0, but 2D data (not as expected) on iteration 1, resulting in
an error.

Appropriately setting `max_frames` of these extra `PluginData` objects
resolves this issue.

(As a side note, you can refer to 27c132c7 to see when these extra
`PluginData` objects were first introduced in the context of
pattern-switching).