Move sign change of angles to 2D case in `ForwardProjectorGpu`
In commit 287794be, the sign change of the angles in
`ForwardProjectorGpu` was introduced. Therefore, in a later commit
97934775, this sign change behaviour was preserved for
- a `ForwardProjectorGpu` plugin not running iteratively (ie, the sign
  change would occur as normal)
- a `ForwardProjectorGpu` plugin running in an iterative loop, but only
  on the 0th iteration (ie, the sign change would occur on iteration 0,
  but on subsequent iterations the sign change would not occur, to
  prevent periodic sign changes across iterations)

The change in 97934775 was to fix a bug which resulted in a
"checkerboard pattern" appearing in the forward projections that were
generated by `ForwardProjectorGpu`, where horizontal blocks of the
forward projections were misaligned.

However, this bug has appeared once more, and was likely caused by
recently changing the `ForwardProjectorGpu` plugin to primarily process
3D blocks rather than 2D slices.

From this, it was concluded that the solution to fixing the bug this
time around was to move the sign change of the angles to the 2D data
case in `ForwardProjectorGpu`.