Begin implementing tomopy's iterative alignment approach
This approach involves transforming the projections with the shifts that
are calculated in the Projection2dAlignment plugin, and then having
TomobarRecon3d perform reconstruction from these shifted projections.

During implementation of this alternative approach, a bug was found in
the iterative functionality where metadata was failing to be passed from
the second output dataset of Projection2dAlignment to the input of
TomobarRecon3d (in particular, the `centre_of_rotation`). This issue has
been resolved in this change as well.

The error in the shifts reduces as more iterations are performed, but
further improvements can still be made to this approach to iterative
alignment.