Add preview info of shifted projs dataset to experimental metadata
The VoCentering plugin requires its input dataset to have preview
information stored in the experimental metadata. Any dataset that is
loaded by a loader plugin will automatically have this done, and thus
will work fine as input to the VoCentering plugin.

However, there are cases where running a non-loader plugin can generate
data that would be desired to pass to VoCentering. One such case is in
the context of iterative alignment in Savu, when Projection2dAlignment
is shifting the projections and generates an output dataset containing
the shifted projections. As this dataset wasn't created by a loader
plugin, it does not have preview information stored in the experimental
metadata, and thus will cause the VoCentering plugin to fail when passed
to it.

By using the same approach that loaders use to add this preview
information to the experimental metadata (calling
`Preview.set_preview()` with the keyword arg `load=True`), the shifted
projections dataset that is generated by Projection2dAlignment can now
be processed by VoCentering without error.