Don't mark output datasets of VoCentering as "to remove" when in a loop
Similar to 6c9625c6, in order for the output datasets of VoCentering to
be reused and updated in subsequent iterations, the associated hdf5
files need to be kept open. However, marking an output dataset as "to
remove" will close the associated hdf5 file when the plugin has finished
its processing. Therefore, when VoCentering is in an iterative loop, it
will now not mark its output datasets like this.