Move job params to be hardcoded in `savu_mpijob.sh`
The only exception is `--output`, because the filepath of the output
file `savu.o*` needs to be dynamic, and is easier to pass as a flag.

A consequence of moving dynamically calculated job parameters from
`savu_launcher.sh` to be hardcoded in `savu_mpijob.sh` is that many
variables in `savu_launcher.sh` are longer necessary (such as
`cpus_per_node`, `cpus_to_use_per_node`, `nNodes`, etc).