Remove copy of process list file to log folder in `savu_launcher.sh`
The way that this SLURM-migrated version of `savu_launcher.sh` is
intended to be run is by first remotely logging onto `wilson`, and then
running the launcher script via the `savu_mpi` alias. Therefore, all
commands in `savu_launcher.sh` are run on `wilson`.

In particular, there is a `cp` command that is run in
`savu_launcher.sh`, seemingly wanting to copy the original prcoess list
file to the log folder associated with the Savu job. However, it is
possible that there exist directories on compute nodes which are not
available on `wilson` in which a process list file may reside (for
example, currently `/mnt/gpfs03` is available on Hamilton nodes, but not
on `wilson`).

If a Savu job is submitted when the process list file is "visible" on
compute nodes but not on `wilson`, this causes the `cp` command to fail,
and eventually cascades down to cause the Savu run to fail.

After some observations of what files are copied by `savu_launcher.sh`
vs. by Savu itself, and some inspection of the Savu source code, it
appears that Savu itself also copies the process list file to the log
folder (seemingly also taking into account if more than one MPI process
exists, to make sure that only one process attempts to do the file-save
operation?).

Therefore, it appeared unnecessary for `savu_launcher.sh` to do the `cp`
at all, and after removing it and checking if the process list file was
copied to the run folder or not, it was indeed still copied without
issue.

Hence, this change has removed the copy of the process list file from
`savu_launcher.sh`, while still preserving the same behaviour as before.